Altered GUI to generate tracks.
Previously, the GUI had tracks/track handles built in manually in QT Designer. With this commit, I have built a track class to generate the track objects using C++ so that tracks can be dynamically added and removed. In addition, this will allow for easy access to the fields of the tracks to do operations on them, such as tracking the volume of the slider and creating the visualization of the audio in the graphics area.